+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -6,9 +6,7 @@ import FlutterMacOS
 import Foundation
 
 import file_picker
-import flutter_bin
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   FilePickerPlugin.register(with: registry.registrar(forPlugin: "FilePickerPlugin"))
-  FlutterBinPlugin.register(with: registry.registrar(forPlugin: "FlutterBinPlugin"))
 }